requirements.txt
requests==2.28.2
Flask==2.2.3
Flask-PyMongo~=2.3.0
cryptography==39.0.2
gunicorn~=20.1.0